backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# File handling
aiofiles==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment
python-dotenv==1.0.0

# Image Processing & OCR (for future medication photo processing)
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.25.2

# Data Processing (for analytics and reporting)
pandas==2.1.3

# Caching & Background Tasks (for future scalability)
redis==5.0.1
celery==5.3.4

# Additional utilities
email-validator==2.1.0
jinja2==3.1.2
